+++ packages/web/src/app/(public)/page.tsx
@@ -109,7 +109,7 @@ export default function LoginPage() {
         </div>
 
         {/* Login Card */}
-        <Card className='py-8 dark:backdrop-blur-sm bg-card/80 dark:bg-background/80'>
+        <Card className='py-2 dark:backdrop-blur-sm bg-card/80 dark:bg-background/80'>
           <CardHeader>
             <CardTitle className='text-xl'>Sign in to your account</CardTitle>
           </CardHeader>